+++ src/backend/core/tests/documents/test_api_documents_media_auth.py
@@ -6,7 +6,6 @@ from io import BytesIO
 from urllib.parse import urlparse
 from uuid import uuid4
 
-from django.conf import settings
 from django.core.files.storage import default_storage
 from django.utils import timezone
 
@@ -37,7 +36,7 @@ def test_api_documents_media_auth_unkown_document():
     assert models.Document.objects.exists() is False
 
 
-def test_api_documents_media_auth_anonymous_public():
+def test_api_documents_media_auth_anonymous_public(settings):
     """Anonymous users should be able to retrieve attachments linked to a public document"""
     document_id = uuid4()
     filename = f"{uuid4()!s}.jpg"
@@ -139,7 +138,7 @@ def test_api_documents_media_auth_anonymous_authenticated_or_restricted(reach):
     assert "Authorization" not in response
 
 
-def test_api_documents_media_auth_anonymous_attachments():
+def test_api_documents_media_auth_anonymous_attachments(settings):
     """
     Declaring a media key as original attachment on a document to which
     a user has access should give them access to the attachment file
@@ -202,7 +201,9 @@ def test_api_documents_media_auth_anonymous_attachments():
 
 
 @pytest.mark.parametrize("reach", ["public", "authenticated"])
-def test_api_documents_media_auth_authenticated_public_or_authenticated(reach):
+def test_api_documents_media_auth_authenticated_public_or_authenticated(
+    reach, settings
+):
     """
     Authenticated users who are not related to a document should be able to retrieve
     attachments related to a document with public or authenticated link reach.
@@ -284,7 +285,7 @@ def test_api_documents_media_auth_authenticated_restricted():
 
 
 @pytest.mark.parametrize("via", VIA)
-def test_api_documents_media_auth_related(via, mock_user_teams):
+def test_api_documents_media_auth_related(via, mock_user_teams, settings):
     """
     Users who have a specific access to a document, whatever the role, should be able to
     retrieve related attachments.
@@ -368,7 +369,7 @@ def test_api_documents_media_auth_not_ready_status():
     assert response.status_code == 403
 
 
-def test_api_documents_media_auth_missing_status_metadata():
+def test_api_documents_media_auth_missing_status_metadata(settings):
     """Attachments without status metadata should be considered as ready"""
     document_id = uuid4()
     filename = f"{uuid4()!s}.jpg"
@@ -387,6 +388,54 @@ def test_api_documents_media_auth_missing_status_metadata():
     with freeze_time(now):
         response = APIClient().get(
             "/api/v1.0/documents/media-auth/", HTTP_X_ORIGINAL_URL=original_url
+        )
+
+    assert response.status_code == 200
+
+    authorization = response["Authorization"]
+    assert "AWS4-HMAC-SHA256 Credential=" in authorization
+    assert (
+        "SignedHeaders=host;x-amz-content-sha256;x-amz-date, Signature="
+        in authorization
+    )
+    assert response["X-Amz-Date"] == now.strftime("%Y%m%dT%H%M%SZ")
+
+    s3_url = urlparse(settings.AWS_S3_ENDPOINT_URL)
+    file_url = f"{settings.AWS_S3_ENDPOINT_URL:s}/impress-media-storage/{key:s}"
+    response = requests.get(
+        file_url,
+        headers={
+            "authorization": authorization,
+            "x-amz-date": response["x-amz-date"],
+            "x-amz-content-sha256": response["x-amz-content-sha256"],
+            "Host": f"{s3_url.hostname:s}:{s3_url.port:d}",
+        },
+        timeout=1,
+    )
+    assert response.content.decode("utf-8") == "my prose"
+
+
+def test_api_documents_media_auth_anonymous_public_custom_origin_header(settings):
+    """Changing the setting MEDIA_AUTH_ORIGINAL_URL_HEADER to match other header should work"""
+    settings.MEDIA_AUTH_ORIGINAL_URL_HEADER = "HTTP_X_FORWARDED_URI"
+    document_id = uuid4()
+    filename = f"{uuid4()!s}.jpg"
+    key = f"{document_id!s}/attachments/{filename:s}"
+    default_storage.connection.meta.client.put_object(
+        Bucket=default_storage.bucket_name,
+        Key=key,
+        Body=BytesIO(b"my prose"),
+        ContentType="text/plain",
+        Metadata={"status": DocumentAttachmentStatus.READY},
+    )
+
+    factories.DocumentFactory(id=document_id, link_reach="public", attachments=[key])
+
+    original_url = f"http://localhost/media/{key:s}"
+    now = timezone.now()
+    with freeze_time(now):
+        response = APIClient().get(
+            "/api/v1.0/documents/media-auth/", HTTP_X_FORWARDED_URI=original_url
         )
 
     assert response.status_code == 200